requirements.txt
requests==2.31.0
urllib3==1.26.5
certifi==2024.6.2
seleniumbase==4.27.0
